refactor: change button-primary package name to button
