+++ packages/article-list/package.json
@@ -64,7 +64,7 @@
   },
   "dependencies": {
     "@times-components/article-summary": "0.25.5",
-    "@times-components/button-primary": "0.0.1",
+    "@times-components/button": "0.0.1",
     "@times-components/card": "0.30.5",
     "@times-components/error-view": "0.12.5",
     "@times-components/image": "1.19.5",
