--- packages/button-primary/__tests__/android/button-primary.test.js
@@ -1,5 +1,0 @@
-import shared from "../shared";
-
-describe("ButtonPrimary tests on android", () => {
-  shared();
-});
