+++ packages/button/__tests__/shared.js
@@ -1,13 +1,13 @@
 import "react-native";
 import React from "react";
 import { shallow } from "enzyme";
-import ButtonPrimary from "../src/button-primary";
+import Button from "../src/button";
 
 export default () => {
   it("should render the button correctly", () => {
     const onPressMock = () => {};
     const wrapper = shallow(
-      <ButtonPrimary onPress={onPressMock} title="test button" />
+      <Button onPress={onPressMock} title="test button" />
     );
 
     expect(wrapper).toMatchSnapshot("1. Render the button");
@@ -16,7 +16,7 @@ export default () => {
   it("should handle the onPress event", () => {
     const onPressMock = jest.fn();
     const wrapper = shallow(
-      <ButtonPrimary onPress={onPressMock} title="test button" />
+      <Button onPress={onPressMock} title="test button" />
     );
 
     wrapper.simulate("press");
