+++ packages/button/__tests__/web/button.web.test.js
@@ -1,0 +1,5 @@
+import shared from "../shared";
+
+describe("Button tests on web", () => {
+  shared();
+});
