+++ packages/button/button.stories.js
@@ -1,4 +1,4 @@
 import { showcaseConverter } from "@times-components/storybook";
-import showcase from "./button-primary.showcase";
+import showcase from "./button.showcase";
 
 showcaseConverter(module, showcase);
