+++ packages/button/package.json
@@ -1,9 +1,9 @@
 {
-  "name": "@times-components/button-primary",
+  "name": "@times-components/button",
   "version": "0.0.1",
-  "description": "Primary button component",
-  "main": "dist/button-primary",
-  "dev": "src/button-primary",
+  "description": "Button component",
+  "main": "dist/button",
+  "dev": "src/button",
   "scripts": {
     "fmt": "prettier --write '**/*.*'",
     "prettier:diff": "prettier --list-different '**/*.*'",
@@ -25,7 +25,7 @@
     "react",
     "native",
     "web",
-    "button-primary",
+    "button",
     "component"
   ],
   "license": "BSD-3-Clause",
