+++ packages/styleguide/src/fonts/font-sizes.android.js
@@ -2,7 +2,7 @@ import sharedFontSizes from "./font-sizes-base";
 
 const fontSizes = {
   ...sharedFontSizes,
-  buttonPrimary: 14
+  button: 14
 };
 
 export default fontSizes;
